pyproject.toml
[tool.poetry]
name = "cousewok-7"
version = "0.1.0"
description = ""
authors = ["ryfe <denis.koptelev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
drf-yasg = "^1.21.7"
redis = "^5.0.5"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"
flake8 = "^7.0.0"
coverage = "^7.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
